requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter>=24.2
djangorestframework-simplejwt>=5.3
drf-yasg>=1.21
psycopg2-binary>=2.9
django-redis>=5.4
gunicorn>=21.2


